pyproject.toml
[project]
name = "pytorch-playground"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy (==2.2.2)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "torch (>=2.2.0,<3.0.0)",
    "scikit-learn (>=1.3.0,<2.0.0)",
    "torchvision (>=0.17.0,<1.0.0)",
    "pillow (>=10.3.0,<11.0.0)",
    "timm (>=1.0.20,<2.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "streamlit (>=1.50.0,<2.0.0)",
    "opencv-python-headless (>=4.12.0.88,<5.0.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
